cleaned up test examples
