+++ Tests/SwiftletModelTests/Models/Attachment.swift
@@ -27,24 +27,4 @@ extension Schema.V1 {
     }
 }
 
-extension Schema.V1 {
-
-    @EntityModel
-    struct GenericAttachment<T: Codable & Sendable & Hashable>: Codable, Sendable {
-        @HashIndex<Self>(\.kind) private var kindIndex
-        
-        let id: String
-        var kind: T
-        
-        @Relationship
-        var message: Message? = .none
-
-        enum Kind: Codable, Hashable {
-            case image(url: URL)
-            case video(url: URL)
-            case file(url: URL)
-            case generic(T)
-        }
-    }
-}
